feat: use-navbar hook
